feat: Refactor component structure and update imports
This commit refactors the component structure by moving common components into more specific directories and updates the corresponding imports.

-   Moved `StyleGrid`, `ProductGrid`, and `FeaturedStyleGrid` to `ui/screens/product/components/grid/`.
-   Moved `StyleCard` and `ProductCard` to `ui/screens/product/components/cards/`.
-   Moved `ProductSectionsList` and `SectionView` to `ui/screens/product/components/sections/`.
-   Moved `ProductFooter` to `ui/screens/product/components/footer/`.
-   Moved `ProductHorizontalList` to `ui/screens/product/components/lists/`.
-   Moved `BannerSlider` to `ui/screens/product/components/banners/`.
-   Removed `Footer.kt`.
-   Updated import statements in affected files to reflect the new component locations.
