+++ app/src/main/java/com/example/stylefeed/ui/screens/product/components/grid/FeaturedStyleGrid.kt
@@ -1,4 +1,4 @@
-package com.example.stylefeed.ui.common.grid
+package com.example.stylefeed.ui.screens.product.components.grid
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -8,7 +8,6 @@ import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -17,7 +16,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.example.stylefeed.domain.model.Style
-import com.example.stylefeed.ui.common.StyleCard
+import com.example.stylefeed.ui.screens.product.components.cards.StyleCard
 
 @Composable
 fun FeaturedStyleGrid(
