+++ app/src/main/java/com/example/stylefeed/ui/screens/product/components/sections/ProductSectionView.kt
@@ -1,4 +1,4 @@
-package com.example.stylefeed.ui.screens.product.components
+package com.example.stylefeed.ui.screens.product.components.sections
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -8,11 +8,11 @@ import com.example.stylefeed.domain.model.Content
 import com.example.stylefeed.domain.model.FooterType
 import com.example.stylefeed.domain.model.SectionState
 import com.example.stylefeed.domain.model.imageAspectRatio
-import com.example.stylefeed.ui.common.BannerSlider
-import com.example.stylefeed.ui.common.Footer
-import com.example.stylefeed.ui.common.ProductHorizontalList
-import com.example.stylefeed.ui.common.grid.ProductGrid
-import com.example.stylefeed.ui.common.grid.StyleGrid
+import com.example.stylefeed.ui.screens.product.components.banners.BannerSlider
+import com.example.stylefeed.ui.screens.product.components.lists.ProductHorizontalList
+import com.example.stylefeed.ui.screens.product.components.grid.ProductGrid
+import com.example.stylefeed.ui.screens.product.components.grid.StyleGrid
+import com.example.stylefeed.ui.screens.product.components.footer.ProductFooter
 
 @Composable
 fun SectionView(
